+++ src/Phan/Language/Type/NativeType.php
@@ -125,6 +125,36 @@ abstract class NativeType extends Type
             ?? parent::canCastToNonNullableType($type);
     }
 
+    /**
+     * @return bool
+     * True if this Type can be cast to the given Type
+     * cleanly
+     */
+    protected function canCastToNonNullableTypeWithoutConfig(Type $type) : bool
+    {
+        // Anything can cast to mixed or ?mixed
+        // Not much of a distinction in nullable mixed, except to emphasize in comments that it definitely can be null.
+        // MixedType overrides the canCastTo*Type methods to always return true.
+        if ($type instanceof MixedType) {
+            return true;
+        }
+
+        if (!($type instanceof NativeType)
+            || $this instanceof GenericArrayType
+            || $type instanceof GenericArrayType
+        ) {
+            return parent::canCastToNonNullableTypeWithoutConfig($type);
+        }
+
+        static $matrix;
+        if ($matrix === null) {
+            $matrix = self::initializeTypeCastingMatrix();
+        }
+
+        return $matrix[$this->getName()][$type->getName()]
+            ?? parent::canCastToNonNullableTypeWithoutConfig($type);
+    }
+
     protected function isSubtypeOfNonNullableType(Type $type) : bool
     {
         // Anything is a subtype of mixed or ?mixed